refactor(core): migrate rate limiting and numeric types to improve precision
- Changed rate_limit type from f64 to u32 and semantics from milliseconds
  to requests per second
- Updated all exchange implementations to use new rate_limit type
- Migrated BidAsk and MarkPrice structs to use Decimal for financial
  calculations
- Replaced f64-based arithmetic with Decimal operations for better
  precision
- Updated tests to use decimal macros and assertions
- Modified spread and percentage calculations to use Decimal math
- Adjusted rate limiter configuration to match new semantic
- Updated documentation examples to reflect Decimal usage
- Fixed property tests to use exact equality instead of floating point
  comparisons

+++ ccxt-exchanges/src/hyperliquid/exchange_impl.rs
@@ -88,8 +88,8 @@ impl Exchange for HyperLiquid {
         ]
     }
 
-    fn rate_limit(&self) -> f64 {
-        100.0
+    fn rate_limit(&self) -> u32 {
+        100
     }
 
     // ==================== Market Data (Public API) ====================
